deno.
yep. This was fun.

It's still a bit funky to paste in the html file here and the other
things generated in dist/. I might jump into that more at some point.
Holler if you have any thoughts.

TypeScript has some definitions of what MediaTrackCapabilities looks
like, but the browsers don't agree with TypeScript's definition. This is
why I set up the custom interface and Capability type, and used the
`as`. There may be a worthy patch to make to get this lined up.